Reorg of code ready for testing
