+++ src/doorman/Misc.hpp
